+++ src/core/components/ui/index.ts
@@ -8,5 +8,6 @@ export * from "./card";
 export * from "./label";
 export * from "./textarea";
 export * from "./native-select";
+export * from "./pagination";
 export * from "./select";
 export * from "./skeleton";